requirements.txt
Flask==2.2.3
requests==2.22.0
requests-oauthlib==1.3.0
pymysql
itsdangerous==2.0.1
flask-sqlalchemy==3.0.3
gunicorn
flask-session
cachelib
